Corrected all methods that search by id, and removed the option to update a sale

+++ clothing-store/src/main/java/app/clotheStore/service/EmployeeService.java
@@ -2,6 +2,7 @@ package app.clotheStore.service;
 
 import app.clotheStore.entity.Employee;
 import app.clotheStore.repository.EmployeeRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -23,16 +24,24 @@ public class EmployeeService {
     // PUT
     // Update an Employee
     public String update(Employee employee, Long id){
-        employee.setId(id);
-        employeeRepository.save(employee);
-        return employee.getName() + " successfully updated";
+        if (employeeRepository.existsById(id)){
+            employee.setId(id);
+            employeeRepository.save(employee);
+            return employee.getName() + " successfully updated";
+        } else {
+            throw new EntityNotFoundException("Employee with ID: " + id + " not found");
+        }
     }
 
     // DELETE
     //  an Employee
     public String delete(Long id){
-        this.employeeRepository.deleteById(id);
-        return "Employee with id: " + id + " deleted";
+        if (employeeRepository.existsById(id)) {
+            this.employeeRepository.deleteById(id);
+            return "Employee with id: " + id + " deleted";
+        } else {
+            throw new EntityNotFoundException("Employee with ID: " + id + " not found");
+        }
     }
 
     // GET
@@ -44,6 +53,15 @@ public class EmployeeService {
     // GET
     // Find Employee by ID
     public Employee findById(Long id){
-        return this.employeeRepository.findById(id).get();
+        if (employeeRepository.existsById(id)){
+            return this.employeeRepository.findById(id).get();
+        } else {
+            throw new EntityNotFoundException("Employee with ID: " + id + " not found");
+        }
+    }
+
+    // Verify Employee existence by ID
+    public boolean existsById(Long id) {
+        return employeeRepository.existsById(id);
     }
 }
